Move markdown/yaml generators to 'detekt-utils'
