+++ detekt-generator/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
     implementation(projects.detektRulesEmpty)
     implementation(projects.detektFormatting)
     implementation(projects.detektCli)
-    implementation(projects.detektTooling)
+    implementation(projects.detektUtils)
     implementation(libs.jcommander)
 
     testImplementation(projects.detektCore)
