+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/printer/RuleConfigurationPrinter.kt
@@ -1,13 +1,13 @@
 package io.gitlab.arturbosch.detekt.generator.printer
 
-import io.github.detekt.tooling.out.bold
-import io.github.detekt.tooling.out.code
-import io.github.detekt.tooling.out.crossOut
-import io.github.detekt.tooling.out.description
-import io.github.detekt.tooling.out.h4
-import io.github.detekt.tooling.out.item
-import io.github.detekt.tooling.out.list
-import io.github.detekt.tooling.out.markdown
+import io.github.detekt.utils.bold
+import io.github.detekt.utils.code
+import io.github.detekt.utils.crossOut
+import io.github.detekt.utils.description
+import io.github.detekt.utils.h4
+import io.github.detekt.utils.item
+import io.github.detekt.utils.list
+import io.github.detekt.utils.markdown
 import io.gitlab.arturbosch.detekt.generator.collection.Configuration
 
 internal object RuleConfigurationPrinter : DocumentationPrinter<List<Configuration>> {
