+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/printer/defaultconfig/RuleSetConfigPrinter.kt
@@ -1,8 +1,8 @@
 package io.gitlab.arturbosch.detekt.generator.printer.defaultconfig
 
-import io.github.detekt.tooling.out.YamlNode
-import io.github.detekt.tooling.out.keyValue
-import io.github.detekt.tooling.out.node
+import io.github.detekt.utils.YamlNode
+import io.github.detekt.utils.keyValue
+import io.github.detekt.utils.node
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.generator.collection.Configuration
 import io.gitlab.arturbosch.detekt.generator.collection.Rule
