+++ detekt-utils/build.gradle.kts
@@ -1,3 +1,7 @@
 plugins {
     id("module")
 }
+
+dependencies {
+    testImplementation(libs.assertj)
+}
