+++ detekt-utils/src/main/kotlin/io/github/detekt/utils/Markdown.kt
@@ -1,6 +1,6 @@
 @file:Suppress("detekt.TooManyFunctions")
 
-package io.github.detekt.tooling.out
+package io.github.detekt.utils
 
 sealed class Markdown(open var content: String = "") {
     fun append(value: String) {
