+++ detekt-utils/src/main/kotlin/io/github/detekt/utils/Yaml.kt
@@ -1,4 +1,4 @@
-package io.github.detekt.tooling.out
+package io.github.detekt.utils
 
 sealed class YML(open val indent: Int = 0, open var content: String = "") {
     fun append(value: String) {
