+++ detekt-utils/src/test/kotlin/io/github/detekt/utils/YamlSpec.kt
@@ -1,4 +1,4 @@
-package io.github.detekt.tooling.out
+package io.github.detekt.utils
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
